+++ datagym/__init__.py
@@ -1,3 +1,7 @@
 from datagym.client import Client   # noqa
 from datagym.importers.coco import Coco   # noqa
 from datagym.exceptions.exceptions import ClientException, APIException   # noqa
+from datagym.models.label_config import LabelConfig # noqa
+from datagym.models.image import Image  # noqa
+from datagym.models.project import Project  # noqa
+from datagym.models.dataset import Dataset  # noqa