+++ datagym/constants.py
@@ -1,1 +1,1 @@
-WARNING_KEYS = ["ex_alreadyexists", "ex_gen_contains"]
+WARNING_KEYS = ["ex_alreadyexists", "ex_gen_contains", "ex_already_contains"]