+++ datagym/endpoints.py
@@ -24,6 +24,8 @@ class Endpoint:
     _get_dataset_by_id: str = "api/v1/dataset/{dataset_id}"
     _import_labels: str = "api/v1/project/{project_id}/prediction"
     _upload_image: str = "/api/v1/image/{dataset_id}/file"
+    _clear_label_config: str = "/api/v1/config/{configId}"
+    _upload_label_config: str = "/api/v1/config/{configId}"
 
     def export_labels(self, project_id: str) -> str:
         return self._export_labels.format(project_id=project_id)
@@ -57,3 +59,9 @@ class Endpoint:
 
     def upload_image(self, dataset_id: str):
         return self._upload_image.format(dataset_id=dataset_id)
+
+    def clear_label_config(self, config_id: str):
+        return self._clear_label_config.format(configId=config_id)
+
+    def upload_label_config(self, config_id: str):
+        return self._upload_label_config.format(configId=config_id)