+++ datagym/models/__init__.py
@@ -1,1 +1,2 @@
 from datagym.models.project import Project, Dataset, Image  # noqa
+from datagym.models.label_config import LabelConfig  # noqa