+++ HISTORY.rst
@@ -4,14 +4,13 @@ History
 
 0.1.0 (2020-03-31)
 ------------------
-
 * First release on PyPI.
 
 
 0.2.0 (2020-04-06)
 ------------------
-
 * Added Import Labels functionality.
+
 
 0.3.0 (2020-04-27)
 ------------------
@@ -19,7 +18,14 @@ History
 * Removed owner from dataset
 * Url and Label upload in mini batches
 
+
 0.4.0 (2020-04-27)
 ------------------
 * Added warnings for non fatal exceptions
 * Added image upload functionality
+
+
+0.5.0 (2020-06-09)
+------------------
+* Added a label config uploader
+* Adjusted the coco uploader to work together with the label config uploader